// elara-backend-v1/src/modules/wardrobe/domain/enums/pattern.enum.ts
export enum Pattern {
  SOLID = 'solid',
  STRIPED = 'striped',
  CHECKED = 'checked',
  PLAID = 'plaid',
  POLKA_DOT = 'polka-dot',
  PRINTED = 'printed',
  FLORAL = 'floral',
  ABSTRACT = 'abstract',
  GEOMETRIC = 'geometric',
  ANIMAL_PRINT = 'animal-print',
  TIE_DYE = 'tie-dye',
  CAMO = 'camo',
}